frontend: Extract AppContainer from App
